Move unnecessary integration tests to functional
